requirements.txt
asgiref==3.2.3
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.5
django-crispy-forms==1.8.1
django-model-utils==4.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==6.2.0
pkgconfig==1.5.1
psycopg2==2.8.4
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
rules==2.2
serpy==0.3.1
six==1.14.0
sqlparse==0.3.0
stripe==2.42.0
urllib3==1.25.3
